isolated workers, fixed risera, new and better div7a, almost working automated endpoint tests, and much more

+++ docker_scripts/down.sh
@@ -1,7 +1,13 @@
 #!/usr/bin/env fish
-function _; or status --is-interactive; or exit 1; end
-set DIR (dirname (readlink -m (status --current-filename))); cd "$DIR"
-set VENV_PATH ~/.local/robust/$DIR/venv
-. $VENV_PATH/bin/activate.fish ;_
+function e; or status --is-interactive; or exit 1; end
 
-docker-compose  -f ../generated_stack_files/last.yml -p robust --compatibility down
+set DIR (dirname (readlink -m (status --current-filename))); cd "$DIR"
+
+set VENV_PATH ./venv
+. $VENV_PATH/bin/activate.fish ;e
+
+if test -e ./../generated_stack_files/last.yml;
+	docker-compose  -f ../generated_stack_files/last.yml -p robust --compatibility down;
+else;
+	echo "./../generated_stack_files/last.yml not found, nothing to do.";
+end
